// frontend/src/features/profile/types/profile.types.ts
export type ProfileResponse = {
  userId: number;
  firstName: string;
  lastName: string;
  email: string;
  enabled: boolean;
  userStatus: string;
  role: string;
  companyId: number | null;
  companyName: string | null;
  companyActive: boolean | null;
  employeeId: number | null;
  employeeFirstName: string | null;
  employeeLastName: string | null;
  maskedJmbg: string | null;
  phoneCode: string | null;
  phoneNumber: string | null;
  employeeEmail: string | null;
  address: string | null;
  cityId: number | null;
  cityName: string | null;
  postalCode: string | null;
  countryId: number | null;
  countryCode: string | null;
  countryName: string | null;
  timezoneId: number | null;
  timezoneName: string | null;
  timezoneDisplayName: string | null;
  position: string | null;
  employmentDate: string | null;
  active: boolean | null;
  primaryWarehouseId: number | null;
  primaryWarehouseName: string | null;
  primaryWarehouseCompanyId: number | null;
  primaryWarehouseCompanyName: string | null;
};